requirements.txt
substrate-interface==1.2.2
discord.py==1.7.3
redis==4.1.4